refactor channel cmds and fix bugs

+++ http/rest/file.go
@@ -31,6 +31,9 @@ func UploadFile(cmd map[string]interface{}) map[string]interface{} {
 	pwd, _ := cmd["EncryptPassword"].(string)
 	url, _ := cmd["Url"].(string)
 	share, _ := cmd["Share"].(bool)
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	err := dsp.DspService.UploadFile(path, desc, cmd["Duration"], cmd["Interval"], cmd["Times"], cmd["Privilege"], cmd["CopyNum"], pwd, url, whitelist, share)
 	if err != nil {
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
@@ -43,6 +46,9 @@ func DeleteFile(cmd map[string]interface{}) map[string]interface{} {
 	fileHash, ok := cmd["Hash"].(string)
 	if !ok {
 		return ResponsePack(dsp.INVALID_PARAMS)
+	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
 	}
 	ret, err := dsp.DspService.DeleteFile(fileHash)
 	if err != nil {
@@ -61,6 +67,9 @@ func DownloadFile(cmd map[string]interface{}) map[string]interface{} {
 	max, ok := cmd["MaxPeerNum"].(float64)
 	if !ok {
 		max = 1
+	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
 	}
 	err := dsp.DspService.DownloadFile(fileHash, url, link, password, uint64(max))
 	if err != nil {
@@ -98,6 +107,9 @@ func GetUploadFiles(cmd map[string]interface{}) map[string]interface{} {
 		if err != nil {
 			return ResponsePackWithErrMsg(dsp.INVALID_PARAMS, err.Error())
 		}
+	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
 	}
 	log.Debugf("cmd :%v, type %d, offset %d limit %d", cmd, fileType, offset, limit)
 	files, err := dsp.DspService.GetUploadFiles(fileType, offset, limit)
@@ -139,7 +151,9 @@ func GetDownloadFiles(cmd map[string]interface{}) map[string]interface{} {
 		}
 	}
 	log.Debugf("cmd :%v, type %d, offset %d limit %d", cmd, fileType, offset, limit)
-
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	fileinfos, err := dsp.DspService.GetDownloadFiles(fileType, offset, limit)
 	if err != nil {
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
@@ -177,6 +191,9 @@ func GetTransferList(cmd map[string]interface{}) map[string]interface{} {
 			return ResponsePackWithErrMsg(dsp.INVALID_PARAMS, err.Error())
 		}
 	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	list := dsp.DspService.GetTransferList(transferType, offset, limit)
 	resp["Result"] = list
 	return resp
@@ -192,6 +209,9 @@ func CalculateUploadFee(cmd map[string]interface{}) map[string]interface{} {
 	path, err := hex.DecodeString(p)
 	if err != nil {
 		return ResponsePack(dsp.INVALID_PARAMS)
+	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
 	}
 	res, derr := dsp.DspService.CalculateUploadFee(string(path), cmd["Duration"], cmd["Interval"], cmd["Times"], cmd["CopyNum"], cmd["WhiteList"])
 	if derr != nil {
@@ -212,6 +232,9 @@ func GetDownloadFileInfo(cmd map[string]interface{}) map[string]interface{} {
 	if err != nil {
 		return ResponsePack(dsp.INVALID_PARAMS)
 	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	info, derr := dsp.DspService.GetDownloadFileInfo(string(realUrl))
 	if derr != nil {
 		return ResponsePackWithErrMsg(derr.Code, derr.Error.Error())
@@ -231,6 +254,9 @@ func EncryptFile(cmd map[string]interface{}) map[string]interface{} {
 	if !ok || len(password) == 0 {
 		return ResponsePack(dsp.INVALID_PARAMS)
 	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	err := dsp.DspService.EncryptFile(string(path), password)
 	if err != nil {
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
@@ -248,6 +274,9 @@ func DecryptFile(cmd map[string]interface{}) map[string]interface{} {
 	password, ok := cmd["Password"].(string)
 	if !ok || len(password) == 0 {
 		return ResponsePack(dsp.INVALID_PARAMS)
+	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
 	}
 	err := dsp.DspService.DecryptFile(string(path), password)
 	if err != nil {
@@ -283,6 +312,9 @@ func GetFileShareIncome(cmd map[string]interface{}) map[string]interface{} {
 	if err != nil {
 		return ResponsePack(dsp.INVALID_PARAMS)
 	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	ret, derr := dsp.DspService.GetFileShareIncome(begin, end, offset, limit)
 	if derr != nil {
 		return ResponsePackWithErrMsg(derr.Code, derr.Error.Error())
@@ -295,6 +327,9 @@ func GetFileShareRevenue(cmd map[string]interface{}) map[string]interface{} {
 	log.Debugf("GetDownloadFileInfo cmd:%v", cmd)
 	resp := ResponsePack(dsp.SUCCESS)
 	ret := make(map[string]interface{}, 0)
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	revenue, err := dsp.DspService.GetFileRevene()
 	if err != nil {
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
@@ -358,6 +393,9 @@ func WhiteListOperate(cmd map[string]interface{}) map[string]interface{} {
 		})
 	}
 	log.Debugf("fileHash %v, op %v, list %v", fileHash, op, whitelist)
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	tx, err := dsp.DspService.WhiteListOperation(fileHash, uint64(op), whitelist)
 	if err != nil {
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
@@ -373,6 +411,9 @@ func GetFileWhiteList(cmd map[string]interface{}) map[string]interface{} {
 	if !ok || len(fileHash) == 0 {
 		return ResponsePack(dsp.INVALID_PARAMS)
 	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	list, err := dsp.DspService.GetWhitelist(fileHash)
 	if err != nil {
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
@@ -386,6 +427,9 @@ func GetUserSpace(cmd map[string]interface{}) map[string]interface{} {
 	addr, ok := cmd["Addr"].(string)
 	if !ok {
 		return ResponsePack(dsp.INVALID_PARAMS)
+	}
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
 	}
 	userspace, err := dsp.DspService.GetUserSpace(addr)
 	if err != nil {
@@ -414,12 +458,46 @@ func SetUserSpace(cmd map[string]interface{}) map[string]interface{} {
 		secondOp, _ = secondMap["Type"].(float64)
 	}
 	log.Debugf("size %v %v %v %v", uint64(size), uint64(sizeOp), uint64(second), uint64(secondOp))
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	tx, err := dsp.DspService.SetUserSpace(addr, uint64(size), uint64(sizeOp), uint64(second), uint64(secondOp))
 	if err != nil {
 		log.Errorf("add user space err %s", err)
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
 	}
 	resp["Result"] = tx
+	return resp
+}
+
+func GetUserSpaceCost(cmd map[string]interface{}) map[string]interface{} {
+	resp := ResponsePack(dsp.SUCCESS)
+	log.Debugf("cmd %v , type %T", cmd, cmd["Size"])
+	addr, ok := cmd["Addr"].(string)
+	if !ok {
+		return ResponsePack(dsp.INVALID_PARAMS)
+	}
+	size, sizeOp, second, secondOp := float64(0), float64(0), float64(0), float64(0)
+	sizeMap, ok := cmd["Size"].(map[string]interface{})
+	if ok {
+		size, _ = sizeMap["Value"].(float64)
+		sizeOp, _ = sizeMap["Type"].(float64)
+	}
+	secondMap, ok := cmd["Second"].(map[string]interface{})
+	if ok {
+		second, _ = secondMap["Value"].(float64)
+		secondOp, _ = secondMap["Type"].(float64)
+	}
+	log.Debugf("size %v %v %v %v", uint64(size), uint64(sizeOp), uint64(second), uint64(secondOp))
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
+	cost, err := dsp.DspService.GetUserSpaceCost(addr, uint64(size), uint64(sizeOp), uint64(second), uint64(secondOp))
+	if err != nil {
+		log.Errorf("get user space cost err %s", err)
+		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
+	}
+	resp["Result"] = cost
 	return resp
 }
 
@@ -431,6 +509,9 @@ func GetUserSpaceRecords(cmd map[string]interface{}) map[string]interface{} {
 	}
 	offset, _ := dsp.OptionStrToUint64(cmd["Offset"])
 	limit, _ := dsp.OptionStrToUint64(cmd["Limit"])
+	if dsp.DspService == nil {
+		return ResponsePackWithErrMsg(dsp.NO_ACCOUNT, dsp.ErrMaps[dsp.NO_ACCOUNT].Error())
+	}
 	ret, err := dsp.DspService.GetUserspaceRecords(addr, offset, limit)
 	if err != nil {
 		return ResponsePackWithErrMsg(err.Code, err.Error.Error())
